+++ src/LikelihoodBase.cxx
@@ -68,7 +68,7 @@ int KLFitter::LikelihoodBase::SetPhysicsConstants(KLFitter::PhysicsConstants* ph
 // ---------------------------------------------------------
 int KLFitter::LikelihoodBase::SetInitialParameters(std::vector<double> const& parameters) {
   // check number of parameters
-  if (int(parameters.size()) != NParameters()) {
+  if (parameters.size() != NParameters()) {
     std::cout << "KLFitter::SetInitialPosition(). Length of vector does not equal the number of parameters." << std::endl;
     return 0;
   }
@@ -83,7 +83,7 @@ int KLFitter::LikelihoodBase::SetInitialParameters(std::vector<double> const& pa
 // ---------------------------------------------------------
 int KLFitter::LikelihoodBase::SetInitialParametersNChains(std::vector<double> const& parameters, unsigned int nchains) {
   // check number of parameters
-  if (int(parameters.size()) != NParameters()) {
+  if (parameters.size() != NParameters()) {
     std::cout << "KLFitter::SetInitialPosition(). Length of vector does not equal the number of parameters." << std::endl;
     return 0;
   }
@@ -362,7 +362,7 @@ int KLFitter::LikelihoodBase::SetParametersToCache(int iperm, int nperms) {
     fCachedNormalizationVector.assign(nperms, 0.);
   }
 
-  if ((iperm > (int)fCachedParametersVector.size()) || (iperm > (int)fCachedParameterErrorsVector.size())) {
+  if ((iperm > fCachedParametersVector.size()) || (iperm > fCachedParameterErrorsVector.size())) {
     std::cout << "KLFitter::LikelihoodBase::SetParametersToCache: iperm > size of fCachedParametersVector or fCachedParameterErrorsVector!" << std::endl;
     return 0;
   }
@@ -376,7 +376,7 @@ int KLFitter::LikelihoodBase::SetParametersToCache(int iperm, int nperms) {
   int partner = LHInvariantPermutationPartner(iperm, nperms, switchpar1, switchpar2);
 
   if (partner > iperm) {
-    if (((int)fCachedParametersVector.size() > partner)&&((int)fCachedParameterErrorsVector.size() > partner)) {
+    if ((fCachedParametersVector.size() > partner) && (fCachedParameterErrorsVector.size() > partner)) {
       fCachedParametersVector.at(partner) = BCModel::GetBestFitParameters();
       switchcache = fCachedParametersVector.at(partner).at(switchpar1);
       fCachedParametersVector.at(partner).at(switchpar1) = fCachedParametersVector.at(partner).at(switchpar2);
@@ -399,7 +399,7 @@ int KLFitter::LikelihoodBase::SetParametersToCache(int iperm, int nperms) {
 // ---------------------------------------------------------
 
 int KLFitter::LikelihoodBase::GetParametersFromCache(int iperm) {
-  if (((int)fCachedParametersVector.size() > iperm)&&((int)fCachedParameterErrorsVector.size() > iperm)) {
+  if ((fCachedParametersVector.size() > iperm) && (fCachedParameterErrorsVector.size() > iperm)) {
     fCachedParameters = fCachedParametersVector.at(iperm);
     fCachedParameterErrors = fCachedParameterErrorsVector.at(iperm);
     fCachedNormalization = fCachedNormalizationVector.at(iperm);
